+++ extension/language-server/src/tokens/error.ts
@@ -1,3 +1,0 @@
-export const onError = (...args: any[]) => {
-  console.error(args)
-}