+++ extension/ts-plugin/index.ts
@@ -1,2 +1,0 @@
-// @ts-expect-error
-export * from './src'