+++ extension/vscode/src/server.ts
@@ -1,1 +1,0 @@
-import '@pandacss/language-server'