+++ extension/vsix-builder/src/cli-default.ts
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-import { main } from './cli-main'
-
-main().catch((err) => {
-  console.error(err)
-})