+++ extension/vsix-builder/src/vsce/parse-semver.d.ts
@@ -1,9 +1,0 @@
-declare module 'parse-semver' {
-	interface Result {
-		readonly name: string;
-		readonly version: string;
-	}
-	module parseSemver {}
-	function parseSemver(input: string): Result;
-	export = parseSemver;
-}