+++ packages/parser/__tests__/output.test.ts
@@ -890,7 +890,7 @@ describe('extract to css output pipeline', () => {
         {
           "data": [
             {
-              " & > strong": {
+              "& > strong": {
                 "color": "hotpink",
               },
               "color": "lightgreen",
@@ -902,14 +902,14 @@ describe('extract to css output pipeline', () => {
         {
           "data": [
             {
-              " &:hover": {
-                " &:disabled": {
+              "&:hover": {
+                "&:disabled": {
                   "filter": "brightness(1)",
                 },
                 "filter": "brightness(0.85)",
               },
               "@media (min-width: 768px)": {
-                " &:disabled": {
+                "&:disabled": {
                   "filter": "brightness(1)",
                 },
                 "padding": "1rem 0",
@@ -3057,6 +3057,77 @@ describe('extract to css output pipeline', () => {
     `)
   })
 
+  test('array syntax within config recipes', () => {
+    const code = `
+    import { css } from ".panda/css"
+    import { card } from ".panda/recipes"
+
+    export default function Page() {
+      return (
+        <>
+          <div className={cx(card(), css({ fontSize: [2, 5] }))}>Click me!</div>
+        </>
+      )
+    }
+
+     `
+    const result = run(code, {
+      theme: {
+        extend: {
+          recipes: {
+            card: {
+              className: 'card',
+              base: {
+                color: ['blue', 'red'],
+              },
+              variants: {
+                size: {
+                  sm: {
+                    borderRadius: ['sm'],
+                    padding: ['2'],
+                    margin: '4',
+                  },
+                },
+              },
+              defaultVariants: { size: 'sm' },
+            },
+          },
+        },
+      },
+    })
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {
+              "fontSize": [
+                2,
+                5,
+              ],
+            },
+          ],
+          "name": "css",
+          "type": "object",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer utilities {
+        .fs_2 {
+          font-size: 2px;
+        }
+
+        @media screen and (width >= 40em) {
+          .sm\\\\:fs_5 {
+            font-size: 5px;
+          }
+        }
+      }
+      "
+    `)
+  })
+
   test('compositions - textStyle', () => {
     const code = `
         import { css } from "styled-system/jsx"
@@ -3079,6 +3150,7 @@ describe('extract to css output pipeline', () => {
         },
       ]
     `)
+
     expect(result.css).toMatchInlineSnapshot(`
       "@layer utilities {
         @layer compositions {
@@ -3090,6 +3162,324 @@ describe('extract to css output pipeline', () => {
 
         .text_red {
           color: red;
+        }
+      }
+      "
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer utilities {
+        @layer compositions {
+          .textStyle_headline\\\\.h1 {
+            font-size: 2rem;
+            font-weight: var(--font-weights-bold);
+          }
+        }
+
+        .text_red {
+          color: red;
+        }
+      }
+      "
+    `)
+  })
+
+  test('nested outdir + tsconfig.compilerOptions.baseUrl importMap behaviour', () => {
+    const code = `
+    import { css } from "../styled-system/css";
+    import { container } from "../styled-system/patterns";
+
+    export default function App() {
+      return (
+        <div
+          className={container({
+            page: "A4",
+            width: {
+              _print: "210mm",
+            },
+            height: {
+              _print: "297mm",
+              base: "600px",
+            },
+            display: "flex",
+            margin: "auto",
+            flexDir: {
+              _print: "row",
+              base: "column",
+              sm: "row",
+            },
+          })}
+        >
+          <div className={css({ flex: 2 })}>aaa</div>
+          <div className={css({ flex: 1 })}>bbb</div>
+        </div>
+      );
+    }
+
+     `
+    const result = run(code, { outdir: 'src/styled-system', cwd: 'app' }, { compilerOptions: { baseUrl: 'app/src' } })
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {
+              "display": "flex",
+              "flexDir": {
+                "_print": "row",
+                "base": "column",
+                "sm": "row",
+              },
+              "height": {
+                "_print": "297mm",
+                "base": "600px",
+              },
+              "margin": "auto",
+              "page": "A4",
+              "width": {
+                "_print": "210mm",
+              },
+            },
+          ],
+          "name": "container",
+          "type": "pattern",
+        },
+        {
+          "data": [
+            {
+              "flex": 2,
+            },
+          ],
+          "name": "css",
+          "type": "object",
+        },
+        {
+          "data": [
+            {
+              "flex": 1,
+            },
+          ],
+          "name": "css",
+          "type": "object",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer utilities {
+        .pos_relative {
+          position: relative;
+        }
+
+        .max-w_8xl {
+          max-width: var(--sizes-8xl);
+        }
+
+        .mx_auto {
+          margin-inline: auto;
+        }
+
+        .px_4 {
+          padding-inline: var(--spacing-4);
+        }
+
+        .page_A4 {
+          page: A4;
+        }
+
+        .h_600px {
+          height: 600px;
+        }
+
+        .d_flex {
+          display: flex;
+        }
+
+        .m_auto {
+          margin: auto;
+        }
+
+        .flex_column {
+          flex-direction: column;
+        }
+
+        .flex_2 {
+          flex: 2;
+        }
+
+        .flex_1 {
+          flex: 1;
+        }
+
+        @media screen and (width >= 40em) {
+          .sm\\\\:flex_row {
+            flex-direction: row;
+          }
+        }
+
+        @media screen and (width >= 48em) {
+          .md\\\\:px_6 {
+            padding-inline: var(--spacing-6);
+          }
+        }
+
+        @media screen and (width >= 64em) {
+          .lg\\\\:px_8 {
+            padding-inline: var(--spacing-8);
+          }
+        }
+
+        @media print {
+          .print\\\\:w_210mm {
+            width: 210mm;
+          }
+
+          .print\\\\:h_297mm {
+            height: 297mm;
+          }
+
+          .print\\\\:flex_row {
+            flex-direction: row;
+          }
+        }
+      }
+      "
+    `)
+  })
+
+  test('{fn}.raw', () => {
+    const code = `
+    import { css } from ".panda/css";
+    import { buttonStyle } from ".panda/recipes";
+    import { stack } from ".panda/patterns";
+
+    const filePath = String.raw\`C:\\Development\\profile\\aboutme.html\`;
+
+    css.raw({ mx: '3', paddingTop: '4', color: 'amber.100' }, { mx: '10', pt: '6', color: 'blue.950' })
+
+    export default function App() {
+      return (
+        <ButtonStyle rootProps={css.raw({ bg: "red.400" })} />
+      );
+    }
+
+    // recipe in storybook
+    export const Funky: Story = {
+      args: buttonStyle.raw({
+        visual: "funky",
+        shape: "circle",
+        size: "sm",
+      }),
+    };
+
+    // mixed with pattern
+    const stackProps = {
+      sm: stack.raw({ direction: "column" }),
+      md: stack.raw({ direction: "row" })
+    }
+
+    stack(stackProps[props.size]))
+
+     `
+    const result = run(code)
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {
+              "color": "amber.100",
+              "mx": "3",
+              "paddingTop": "4",
+            },
+            {
+              "color": "blue.950",
+              "mx": "10",
+              "pt": "6",
+            },
+          ],
+          "name": "css",
+          "type": "object",
+        },
+        {
+          "data": [
+            {
+              "bg": "red.400",
+            },
+          ],
+          "name": "css",
+          "type": "object",
+        },
+        {
+          "data": [
+            {},
+          ],
+          "name": "ButtonStyle",
+          "type": "jsx-recipe",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer recipes {
+        @layer _base {
+          .buttonStyle {
+            justify-content: center;
+            align-items: center;
+            display: inline-flex;
+          }
+
+          .buttonStyle:is(:hover, [data-hover]) {
+            background-color: var(--colors-red-200);
+          }
+        }
+
+        .buttonStyle--size_md {
+          min-width: 3rem;
+          height: 3rem;
+          padding: 0 .75rem;
+        }
+
+        .buttonStyle--variant_solid {
+          color: var(--colors-white);
+          background-color: #00f;
+        }
+
+        .buttonStyle--variant_solid:is(:hover, [data-hover]) {
+          background-color: #00008b;
+        }
+
+        .buttonStyle--variant_solid[data-disabled] {
+          color: var(--colors-black);
+          background-color: gray;
+        }
+      }
+
+      @layer utilities {
+        .mx_3 {
+          margin-inline: var(--spacing-3);
+        }
+
+        .pt_4 {
+          padding-top: var(--spacing-4);
+        }
+
+        .text_amber\\\\.100 {
+          color: var(--colors-amber-100);
+        }
+
+        .mx_10 {
+          margin-inline: var(--spacing-10);
+        }
+
+        .pt_6 {
+          padding-top: var(--spacing-6);
+        }
+
+        .text_blue\\\\.950 {
+          color: var(--colors-blue-950);
+        }
+
+        .bg_red\\\\.400 {
+          background: var(--colors-red-400);
         }
       }
       "