+++ packages/config/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pandacss/config
 
+## 0.17.5
+
+### Patch Changes
+
+- @pandacss/error@0.17.5
+- @pandacss/logger@0.17.5
+- @pandacss/preset-base@0.17.5
+- @pandacss/preset-panda@0.17.5
+- @pandacss/types@0.17.5
+
 ## 0.17.4
 
 ### Patch Changes