+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pandacss/generator
 
+## 0.17.5
+
+### Patch Changes
+
+- 6718f81b: Fix issue where Solid.js styled factory fails with pattern styles includes a css variable (e.g. Divider)
+- 3ce70c37: Fix issue where cva composition in styled components doens't work as expected.
+- Updated dependencies [a6dfc944]
+  - @pandacss/core@0.17.5
+  - @pandacss/is-valid-prop@0.17.5
+  - @pandacss/logger@0.17.5
+  - @pandacss/shared@0.17.5
+  - @pandacss/token-dictionary@0.17.5
+  - @pandacss/types@0.17.5
+
 ## 0.17.4
 
 ### Patch Changes