+++ packages/is-valid-prop/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @pandacss/is-valid-prop
 
+## 0.17.5
+
 ## 0.17.4
 
 ## 0.17.3